requirements.txt
aioboto3==14.1.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
amazon-transcribe==0.6.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.9.0
asttokens
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
awscrt==0.16.26
boto3==1.37.1
botocore==1.37.1
certifi 
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm 
debugpy 
decorator 
distro==1.9.0
docstring_parser==0.16
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing 
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
importlib_metadata 
ipykernel 
ipython
jedi 
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jupyter_client 
jupyter_core
livekit==1.0.9
livekit-agents==1.1.4
livekit-api==1.0.2
livekit-plugins-aws==1.1.1
livekit-plugins-cartesia==1.1.1
livekit-plugins-deepgram==1.1.1
livekit-plugins-elevenlabs==1.1.1
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.1.1
livekit-plugins-rag==1.0.13
livekit-plugins-silero==1.1.1
livekit-plugins-turn-detector==1.1.1
livekit-protocol==1.0.3
MarkupSafe==3.0.2
matplotlib-inline
mpmath==1.3.0
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.0.2
onnxruntime==1.19.2
openai==1.86.0
packaging==25.0
pandas==2.3.0
parso
pickleshare 
pillow==11.2.1
platformdirs 
prompt_toolkit
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pure_eval
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments 
PyJWT==2.10.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
# pywin32==304
PyYAML==6.0.2
pyzmq 
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
stack_data 
sympy==1.14.0
tokenizers==0.21.1
tornado 
tqdm==4.67.1
traitlets
transformers==4.52.2
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
watchfiles==1.0.5
wcwidth 
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp 
sqlalchemy
psycopg2-binary